--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "liveServer.settings.port": 5503
-}
