+++ addnew.css
@@ -1,36 +1,34 @@
 #addnewbooktitle {
-    text-align: center;
-  }
+  text-align: center;
+}
 
-  input {
-    width: 100%;
-    padding: 10px;
-    margin-bottom: 20px;
-  }
+input {
+  width: 100%;
+  padding: 10px;
+  margin-bottom: 20px;
+}
 
-  form {
-    width: 30%;
-    margin: 0 auto;
-  }
+form {
+  width: 30%;
+  margin: 0 auto;
+}
 
-  .form-class {
-    display: none;
-  }
+.form-class {
+  display: none;
+}
 
-  .hide-form {
-    display: none;
-  }
+.hide-form {
+  display: none;
+}
 
-  #add-btn {
-    background-color: goldenrod;
-  }
+#add-btn {
+  background-color: goldenrod;
+}
 
-  #add-btn:hover {
-    background-color: beige;
-  }
-  
-  #add-btn:active {
-    background-color: orange;
-  }
+#add-btn:hover {
+  background-color: beige;
+}
 
-  
+#add-btn:active {
+  background-color: orange;
+}
