+++ modules/book.js
@@ -1,0 +1,8 @@
+class Book {
+  constructor(id, title, author) {
+    this.id = id;
+    this.title = title;
+    this.author = author;
+  }
+}
+export default Book;
