--- modules/content.js
@@ -1,66 +1,0 @@
-
-class Book {
-  constructor(title, author, showbtn) {
-    this.title = title;
-    this.author = author;
-    this.showbtn = showbtn;
-  }
-}
-class UI {
-  static displayBooks() {
-    const books = Store.getBooks();
-    books.forEach((book) => UI.addBookToList(book));
-  }
-  static addBookToList(book) {
-    const list = document.querySelector('#book-list');
-    const row = document.createElement('tr');
-    row.innerHTML = `
-      <td>${book.title}</td> bn nmgjjbbjf
-      <td>${book.author}</td>
-      <td>${book.showbtn}</td>
-      <td><a href="#" class="btn btn-danger btn-sm delete">X</a></td>
-    `;
-    list.appendChild(row);
-  }
-  static deleteBook(el) {
-    if(el.classList.contains('delete')) {
-      el.parentElement.parentElement.remove();
-    }
-  }
-  static showAlert(message, className) {
-    const div = document.createElement('div');
-    div.className = `alert alert-${className}`;
-    div.appendChild(document.createTextNode(message));
-    const container = document.querySelector('.container');
-    const form = document.querySelector('#book-form');
-    container.insertBefore(div, form);
-    setTimeout(() => document.querySelector('.alert').remove(), 3000);
-  }
-  static clearFields() {
-    document.querySelector('#title').value = '';
-    document.querySelector('#author').value = '';
-    document.querySelector('#showbtn').value = '';
-  }
-}
-
-document.addEventListener('DOMContentLoaded', UI.displayBooks);
-document.querySelector('#book-form').addEventListener('submit', (e) => {
-  e.preventDefault();
-  const title = document.querySelector('#title').value;
-  const author = document.querySelector('#author').value;
-  const showbtn = document.querySelector('#showbtn').value;
-  if(title === '' || author === '' || showbtn === '') {
-    UI.showAlert('please fill the form,thanks!', 'danger');
-  } else {
-    const book = new Book(title, author, showbtn);
-    UI.addBookToList(book);
-    Store.addBook(book);
-    UI.showAlert('Book Added', 'success');
-    UI.clearFields();
-  }
-});
-document.querySelector('#book-list').addEventListener('click', (e) => {
-  UI.deleteBook(e.target);
-  Store.removeBook(e.target.parentElement.previousElementSibling.textContent);
-  UI.showAlert('Book Removed', 'success');
-});
