+++ style.css
@@ -41,11 +41,11 @@ a {
   color: white;
 }
 
- #booklisttitle {
+#booklisttitle {
   text-align: center;
 }
 
- #addnewbooktitle {
+#addnewbooktitle {
   text-align: center;
 }
 
@@ -56,9 +56,7 @@ a {
   border-radius: 16px;
   padding: 25px 20px;
 }
-.current-date {
-  display: none;
-}
+
 .empty {
   justify-content: space-around;
   align-items: center;
@@ -129,20 +127,19 @@ input {
   font-size: 2em;
   font-weight: bold;
   list-style: none;
-  
 }
 
 .hide-contact {
   display: none;
-} 
+}
 
- .hide-form {
+.hide-form {
   display: none;
-} 
+}
 
- .hide-list {
+.hide-list {
   display: none;
-} 
+}
 
 footer {
   text-align: center;
